+++ grain/experimental.py
@@ -37,6 +37,10 @@ from grain._src.python.dataset.dataset import (
 from grain._src.python.dataset.elastic_iterator import ElasticIterator
 from grain._src.python.dataset.sources.parquet_dataset import ParquetIterDataset
 from grain._src.python.dataset.sources.tfrecord_dataset import TFRecordIterDataset
+from grain._src.python.dataset.sources.hf_source import (
+    HFIterableDatasetProtocol,
+    HFIterDataset,
+)
 
 from grain._src.python.dataset.transformations.batch import batch_and_pad
 from grain._src.python.dataset.transformations.cache import CacheIterDataset